+++ ethers-core/src/types/transaction/eip2930.rs
@@ -6,8 +6,9 @@ use open_fastrlp::{
     RlpDecodable as FastRlpDecodable, RlpDecodableWrapper as FastRlpDecodableWrapper,
     RlpEncodable as FastRlpEncodable, RlpEncodableWrapper as FastRlpEncodableWrapper,
 };
-use rlp::{Decodable, RlpStream};
-use rlp_derive::{RlpDecodable, RlpDecodableWrapper, RlpEncodable, RlpEncodableWrapper};
+use rlp::{
+    Decodable, RlpDecodable, RlpDecodableWrapper, RlpEncodable, RlpEncodableWrapper, RlpStream,
+};
 use serde::{Deserialize, Serialize};
 use thiserror::Error;