+++ examples/ethers-wasm/src/lib.rs
@@ -1,8 +1,8 @@
-use crate::utils::SIMPLECONTRACT_BIN;
 use ethers::{
     contract::abigen,
-    prelude::{ContractFactory, Provider, SignerMiddleware},
-    providers::Ws,
+    prelude::{Provider, SignerMiddleware},
+    providers::{Middleware, Ws},
+    signers::Signer,
 };
 use std::sync::Arc;
 use wasm_bindgen::prelude::*;
@@ -10,23 +10,13 @@ use web_sys::console;
 
 pub mod utils;
 
-// When the `wee_alloc` feature is enabled, use `wee_alloc` as the global
-// allocator.
-#[cfg(feature = "wee_alloc")]
-#[global_allocator]
-static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
-
 macro_rules! log {
     ( $( $t:tt )* ) => {
         web_sys::console::log_1(&format!( $( $t )* ).into());
     }
 }
 
-abigen!(
-    SimpleContract,
-    "./abi/contract_abi.json",
-    event_derives(serde::Deserialize, serde::Serialize)
-);
+abigen!(SimpleContract, "./abi/contract.json", derives(serde::Deserialize, serde::Serialize));
 
 #[wasm_bindgen]
 pub async fn deploy() {
@@ -37,23 +27,20 @@ pub async fn deploy() {
         &serde_wasm_bindgen::to_value(&*SIMPLECONTRACT_ABI).unwrap(),
     );
 
-    let wallet = utils::key(0);
-    log!("Wallet: {:?}", wallet);
-
     let endpoint = "ws://127.0.0.1:8545";
     let provider = Provider::<Ws>::connect(endpoint).await.unwrap();
-    let client = Arc::new(SignerMiddleware::new(provider, wallet));
-    log!("Connected to: `{}`", endpoint);
+    log!("Connected to: `{endpoint}`");
 
-    let bytecode = hex::decode(SIMPLECONTRACT_BIN).unwrap();
-    let factory = ContractFactory::new(SIMPLECONTRACT_ABI.clone(), bytecode.into(), client.clone());
+    let chain_id = provider.get_chainid().await.unwrap();
+    let wallet = utils::key(0).with_chain_id(chain_id.as_u64());
+    log!("Wallet: {wallet:?}");
+    let client = Arc::new(SignerMiddleware::new(provider, wallet));
 
     log!("Deploying contract...");
-    let contract = factory.deploy("hello WASM!".to_string()).unwrap().send().await.unwrap();
+    let deploy_tx = SimpleContract::deploy(client.clone(), "hello WASM!".to_string()).unwrap();
+    let contract: SimpleContract<_> = deploy_tx.send().await.unwrap();
     let addr = contract.address();
     log!("Deployed contract with address: {:?}", addr);
-
-    let contract = SimpleContract::new(addr, client.clone());
 
     let value = "bye from WASM!";
     log!("Setting value... `{}`", value);